Fix FILE* left open in from_file
