+++ src/stim/dem/detector_error_model.pybind.cc
@@ -20,6 +20,7 @@
 #include "stim/dem/detector_error_model_instruction.pybind.h"
 #include "stim/dem/detector_error_model_repeat_block.pybind.h"
 #include "stim/dem/detector_error_model_target.pybind.h"
+#include "stim/io/raii_file.h"
 #include "stim/py/base.pybind.h"
 #include "stim/search/search.h"
 
@@ -797,22 +798,16 @@ void pybind_detector_error_model(pybind11::module &m) {
         [](pybind11::object &obj) {
             try {
                 auto path = pybind11::cast<std::string>(obj);
-                FILE *f = fopen(path.data(), "r");
-                if (f == nullptr) {
-                    throw std::invalid_argument("Failed to open " + path);
-                }
-                return DetectorErrorModel::from_file(f);
+                RaiiFile f(path.data(), "r");
+                return DetectorErrorModel::from_file(f.f);
             } catch (pybind11::cast_error &ex) {
             }
 
             auto py_path = pybind11::module::import("pathlib").attr("Path");
             if (pybind11::isinstance(obj, py_path)) {
                 auto path = pybind11::cast<std::string>(pybind11::str(obj));
-                FILE *f = fopen(path.data(), "r");
-                if (f == nullptr) {
-                    throw std::invalid_argument("Failed to open " + path);
-                }
-                return DetectorErrorModel::from_file(f);
+                RaiiFile f(path.data(), "r");
+                return DetectorErrorModel::from_file(f.f);
             }
 
             auto py_text_io_base = pybind11::module::import("io").attr("TextIOBase");
