+++ src/stim/simulators/measurements_to_detection_events.pybind.h
@@ -22,14 +22,6 @@
 #include "stim/circuit/circuit.h"
 #include "stim/mem/simd_bits.h"
 
-struct RaiiFile {
-    FILE *f;
-    RaiiFile(const char *path, const char *mode);
-    RaiiFile(const RaiiFile &other) = delete;
-    RaiiFile(RaiiFile &&other) = delete;
-    ~RaiiFile();
-};
-
 struct CompiledMeasurementsToDetectionEventsConverter {
     const bool skip_reference_sample;
     const stim::simd_bits ref_sample;
